Move certgen/filesystem utilities in an openshift package
