+++ pkg/openshift/certgen/doc.go
@@ -1,0 +1,4 @@
+// Package certgen provides utilities related
+// to generating all the necessary artifacts
+// for an OpenShift deployment.
+package certgen
