+++ pkg/openshift/certgen/templates/master/etc/etcd/etcd.conf
