+++ pkg/openshift/certgen/templates/master/etc/origin/master/master-config.yaml
