+++ pkg/openshift/certgen/templates/master/etc/origin/master/scheduler.json
