+++ pkg/openshift/certgen/templates/master/etc/origin/master/session-secrets.yaml
