+++ pkg/openshift/certgen/templates/master/tmp/ansible/ansible.sh
