+++ pkg/openshift/certgen/templates/master/tmp/ansible/azure-local-master-inventory.yml
