+++ pkg/openshift/certgen/templates/master/tmp/ansible/deploy-router.yml
