+++ pkg/openshift/certgen/templates/master/tmp/ansible/service-catalog.yaml
