+++ pkg/openshift/certgen/templates/master/tmp/bootstrapconfigs/compute-config.yaml
