+++ pkg/openshift/certgen/templates/master/tmp/bootstrapconfigs/infra-config.yaml
