+++ pkg/openshift/certgen/templates/master/tmp/bootstrapconfigs/master-config.yaml
