+++ pkg/openshift/certgen/templates/node/etc/azure/azure.conf
