+++ pkg/openshift/certgen/templates/node/etc/origin/node/node-dnsmasq.conf
