+++ pkg/openshift/certgen/templates/node/etc/origin/node/resolv.conf
