+++ pkg/openshift/certgen/templates/templates.go
