+++ pkg/openshift/doc.go
@@ -1,0 +1,3 @@
+// Package openshift provides utilities related
+// to the OpenShift orchestrator.
+package openshift
